reformat processor/minio.py and move grdc.py to reader dir
