+++ tests/test_aoi_reading.py
@@ -1,7 +1,7 @@
 """
 Author: Wenyu Ouyang
 Date: 2023-10-06 21:17:17
-LastEditTime: 2023-10-10 10:12:04
+LastEditTime: 2023-10-22 16:51:42
 LastEditors: Wenyu Ouyang
 Description: Test reading according to AOI
 FilePath: \hydro_opendata\tests\test_aoi_reading.py
@@ -11,7 +11,7 @@ import os
 import pytest
 import geopandas as gpd
 
-from hydro_opendata.s3api.minio import ERA5LReader, GPMReader, GFSReader
+from hydro_opendata.reader.minio import ERA5LReader, GPMReader, GFSReader
 import numpy as np
